Formatting and bingen code generation
For some reason, that is like that types might be used, bindgen always
seems to insert generated `use` statements for three particular integer
types. We use none of them and not having those types avoids the
requirement of depending on one of libc or std. Thus we fake our own
instance of the type which is a non-public detail.