Run cargo fmt on the code base
This is a nearly automated commit, apart from one manual touch-up. In
`build.rs` one comment had been added after one line of a multi-line of
a conditional control expression. The formatter interprets such comments
as belonging to the body while I'm confident it was meant for the ARM
special case. The comment was thus moved upwards.